# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CreateUserDto": {
				"properties": {
					"personID": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"personID",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseErrorResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"personID": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"personID",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponseDto_UserEntity_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/UserEntity"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponseDto_UserEntity-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"personID": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"personID",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponseDto_string_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePersonDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PersonEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponseDto_PersonEntity_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/PersonEntity"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponseDto_PersonEntity-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PersonEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PersonDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCategoryDto": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponseDto_CategoryEntity_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/CategoryEntity"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponseDto_CategoryEntity-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/CategoryEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateBookDto": {
				"properties": {
					"categoryID": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"author": {
						"type": "string"
					}
				},
				"required": [
					"categoryID",
					"title",
					"author"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookEntity": {
				"properties": {
					"id": {
						"type": "string"
					},
					"categoryID": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"author": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"categoryID",
					"title",
					"author"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponseDto_BookEntity_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/BookEntity"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponseDto_BookEntity-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/BookEntity"
						},
						"type": "array"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"categoryID": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"author": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"categoryID",
					"title",
					"author"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "API Projeto II Documentação",
		"version": "1.0.0",
		"description": "Documentação da API usando Swagger"
	},
	"paths": {
		"/user": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_UserEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_UserEntity-Array_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_UserEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserDto"
							}
						}
					}
				}
			}
		},
		"/user/{userID}": {
			"get": {
				"operationId": "GetUserByID",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_UserEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"202": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_string_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/person": {
			"post": {
				"operationId": "RegisterPerson",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_PersonEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Person"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePersonDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetPersons",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_PersonEntity-Array_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Person"
				],
				"security": [],
				"parameters": []
			},
			"put": {
				"operationId": "UpdatePerson",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_PersonEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Person"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PersonDto"
							}
						}
					}
				}
			}
		},
		"/person/{personID}": {
			"get": {
				"operationId": "GetPersonByID",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_PersonEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Person"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "personID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeletePerson",
				"responses": {
					"202": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_string_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Person"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "personID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/category": {
			"post": {
				"operationId": "RegisterCategory",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_CategoryEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategoryDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_CategoryEntity-Array_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_CategoryEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CategoryDto"
							}
						}
					}
				}
			}
		},
		"/category/{categoryID}": {
			"get": {
				"operationId": "GetCategoryByID",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_CategoryEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"202": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_string_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "categoryID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/book": {
			"post": {
				"operationId": "RegisterBook",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_BookEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Book"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBookDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetBooks",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_BookEntity-Array_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Book"
				],
				"security": [],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateBook",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_BookEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Book"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BookDto"
							}
						}
					}
				}
			}
		},
		"/book/{bookID}": {
			"get": {
				"operationId": "GetBookByID",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_BookEntity_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Book"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteBook",
				"responses": {
					"202": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponseDto_string_"
								}
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseErrorResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Book"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "bookID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Servidor de Desenvolvimento"
		}
	]
}